+++ lms/djangoapps/instructor_task/tests/test_tasks_helper.py
@@ -21,22 +21,6 @@ import tempfile
 import unicodecsv
 from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
-from django.conf import settings
-
-from pytz import UTC
-
-from courseware.courses import get_course_by_id
-from courseware.tests.factories import StudentModuleFactory
-from opaque_keys.edx.keys import CourseKey
-from opaque_keys.edx.locations import Location
-from student.tests.factories import CourseEnrollmentFactory, UserFactory
-from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
-
-TEST_COURSE_ORG = 'edx'
-TEST_COURSE_NAME = 'test_course'
-TEST_COURSE_NUMBER = '1.23x'
-from lms.djangoapps.instructor_task.models import ReportStore
-from student.models import CourseEnrollment
 
 from capa.tests.response_xml_factory import MultipleChoiceResponseXMLFactory
 from certificates.models import CertificateStatuses, GeneratedCertificate
@@ -67,10 +51,6 @@ from xmodule.partitions.partitions import Group, UserPartition
 from lms.djangoapps.instructor_task.models import ReportStore
 from survey.models import SurveyForm, SurveyAnswer
 from lms.djangoapps.instructor_task.tasks_helper import (
-    push_student_responses_to_s3,
-    push_ora2_responses_to_s3,
-    push_course_forums_data_to_s3,
-    push_student_forums_data_to_s3,
     cohort_students_and_upload,
     upload_problem_responses_csv,
     upload_grades_csv,
@@ -107,8 +87,6 @@ class InstructorGradeReportTestCase(TestReportMixin, InstructorTaskCourseTestCas
                 for row in unicodecsv.DictReader(csv_file):
                     if row.get('username') == username:
                         self.assertEqual(row[column_header], expected_cell_content)
-
-TEST_DATA_DIR = settings.COMMON_TEST_DATA_ROOT
 
 
 @ddt.ddt
@@ -913,32 +891,6 @@ class TestProblemReportCohortedContent(TestReportMixin, ContentGroupTestCase, In
 
 
 @ddt.ddt
-class TestReportStore(TestReportMixin, InstructorTaskCourseTestCase):
-    """Tests for the ReportStore model"""
-
-    def setUp(self):
-        super(TestReportStore, self).setUp()
-        self.course = CourseFactory.create()
-
-    def test_delete_report(self):
-        report_store = ReportStore.from_config(config_name='GRADES_DOWNLOAD')
-        task_input = {'features': []}
-
-        links = report_store.links_for(self.course.id)
-        self.assertEquals(len(links), 0)
-        with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task'):
-            upload_students_csv(None, None, self.course.id, task_input, 'calculated')
-        links = report_store.links_for(self.course.id)
-        self.assertEquals(len(links), 1)
-
-        filename = links[0][0]
-        report_store.delete_file(self.course.id, filename)
-
-        links = report_store.links_for(self.course.id)
-        self.assertEquals(len(links), 0)
-
-
-@ddt.ddt
 class TestExecutiveSummaryReport(TestReportMixin, InstructorTaskCourseTestCase):
     """
     Tests that Executive Summary report generation works.
@@ -1268,212 +1220,6 @@ class TestTeamStudentReport(TestReportMixin, InstructorTaskCourseTestCase):
 
         self._generate_and_verify_teams_column(self.student1.username, UNAVAILABLE)
         self._generate_and_verify_teams_column(self.student2.username, team2.name)
-
-
-class TestReponsesReport(TestReportMixin, ModuleStoreTestCase):
-    """
-    Tests that CSV student responses report generation works.
-    """
-
-    def test_unicode(self):
-        self.course = CourseFactory.create()
-        course_key = self.course.id
-        self.problem_location = Location("edX", "unicode_graded", "2012_Fall", "problem", "H1P1")
-
-        self.student = UserFactory(username=u'student\xec')
-        CourseEnrollmentFactory.create(user=self.student, course_id=self.course.id)
-
-        StudentModuleFactory.create(
-            course_id=self.course.id,
-            module_state_key=self.problem_location,
-            student=self.student,
-            grade=0,
-            state=u'{"student_answers":{"fake-problem":"caf\xe9"}}',
-        )
-
-        result = push_student_responses_to_s3(None, None, self.course.id, None, 'generated')
-        self.assertEqual(result, "succeeded")
-
-
-class TestInstructorOra2Report(TestReportMixin, InstructorTaskCourseTestCase):
-    """
-    Tests that ORA2 response report generation works.
-    """
-    def setUp(self):
-        super(TestInstructorOra2Report, self).setUp()
-        self.course = CourseFactory.create(org=TEST_COURSE_ORG,
-                                           number=TEST_COURSE_NUMBER,
-                                           display_name=TEST_COURSE_NAME)
-
-        self.current_task = Mock()
-        self.current_task.update_state = Mock()
-
-    def tearDown(self):
-        if os.path.exists(settings.ORA2_RESPONSES_DOWNLOAD['ROOT_PATH']):
-            shutil.rmtree(settings.ORA2_RESPONSES_DOWNLOAD['ROOT_PATH'])
-
-    def test_report_fails_if_error(self):
-        with patch('lms.djangoapps.instructor_task.tasks_helper.collect_anonymous_ora2_data') as mock_collect_data:
-            mock_collect_data.side_effect = KeyError
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-                mock_current_task.return_value = self.current_task
-
-                self.assertEqual(push_ora2_responses_to_s3(None, None, self.course.id, {'include_email': 'False'}, 'generated'), UPDATE_STATUS_FAILED)
-
-    @patch('lms.djangoapps.instructor_task.tasks_helper.datetime')
-    def test_report_stores_results(self, mock_time):
-        start_time = datetime.now(UTC)
-        mock_time.now.return_value = start_time
-
-        test_header = ['field1', 'field2']
-        test_rows = [['row1_field1', 'row1_field2'], ['row2_field1', 'row2_field2']]
-
-        with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-            mock_current_task.return_value = self.current_task
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper.collect_anonymous_ora2_data') as mock_collect_data:
-                mock_collect_data.return_value = (test_header, test_rows)
-
-                with patch('lms.djangoapps.instructor_task.models.DjangoStorageReportStore.store_rows') as mock_store_rows:
-
-                    timestamp_str = start_time.strftime('%Y-%m-%d-%H%M')
-                    course_id_string = urllib.quote(self.course.id.to_deprecated_string().replace('/', '_'))
-                    filename = u'{}_ORA2_responses_anonymous_{}.csv'.format(course_id_string, timestamp_str)
-                    return_val = push_ora2_responses_to_s3(None, None, self.course.id, {'include_email': 'False'}, 'generated')
-                    self.assertEqual(return_val, UPDATE_STATUS_SUCCEEDED)
-                    mock_store_rows.assert_called_once_with(self.course.id, filename, [test_header] + test_rows)
-
-
-class TestInstructorOra2EmailReport(TestReportMixin, InstructorTaskCourseTestCase):
-    """
-    Tests that ORA2 response report generation works.
-    """
-    def setUp(self):
-        super(TestInstructorOra2EmailReport, self).setUp()
-        self.course = CourseFactory.create(org=TEST_COURSE_ORG,
-                                           number=TEST_COURSE_NUMBER,
-                                           display_name=TEST_COURSE_NAME)
-
-        self.current_task = Mock()
-        self.current_task.update_state = Mock()
-
-    def tearDown(self):
-        if os.path.exists(settings.ORA2_RESPONSES_DOWNLOAD['ROOT_PATH']):
-            shutil.rmtree(settings.ORA2_RESPONSES_DOWNLOAD['ROOT_PATH'])
-
-    def test_report_fails_if_error(self):
-        with patch('lms.djangoapps.instructor_task.tasks_helper.collect_email_ora2_data') as mock_collect_data:
-            mock_collect_data.side_effect = KeyError
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-                mock_current_task.return_value = self.current_task
-
-                self.assertEqual(push_ora2_responses_to_s3(None, None, self.course.id, {'include_email': 'True'}, 'generated'), UPDATE_STATUS_FAILED)
-
-    @patch('lms.djangoapps.instructor_task.tasks_helper.datetime')
-    def test_report_stores_results(self, mock_time):
-        start_time = datetime.now(UTC)
-        mock_time.now.return_value = start_time
-
-        test_header = ['field1', 'field2']
-        test_rows = [['row1_field1', 'row1_field2'], ['row2_field1', 'row2_field2']]
-
-        with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-            mock_current_task.return_value = self.current_task
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper.collect_email_ora2_data') as mock_collect_data:
-                mock_collect_data.return_value = (test_header, test_rows)
-
-                with patch('lms.djangoapps.instructor_task.models.DjangoStorageReportStore.store_rows') as mock_store_rows:
-                    timestamp_str = start_time.strftime('%Y-%m-%d-%H%M')
-                    course_id_string = urllib.quote(self.course.id.to_deprecated_string().replace('/', '_'))
-                    filename = u'{}_ORA2_responses_including_email_{}.csv'.format(course_id_string, timestamp_str)
-                    return_val = push_ora2_responses_to_s3(None, None, self.course.id, {'include_email': 'True'}, 'generated')
-                    self.assertEqual(return_val, UPDATE_STATUS_SUCCEEDED)
-                    mock_store_rows.assert_called_once_with(self.course.id, filename, [test_header] + test_rows)
-
-
-class TestInstructorCourseForumsReport(TestReportMixin, InstructorTaskCourseTestCase):
-    """
-    Tests that course forums usage report generation works.
-    """
-    def setUp(self):
-        super(TestInstructorCourseForumsReport, self).setUp()
-        self.course = CourseFactory.create(org=TEST_COURSE_ORG,
-                                           number=TEST_COURSE_NUMBER,
-                                           display_name=TEST_COURSE_NAME)
-
-        self.current_task = Mock()
-        self.current_task.update_state = Mock()
-
-    def test_report_fails_if_error(self):
-        with patch('lms.djangoapps.instructor_task.tasks_helper.collect_course_forums_data') as mock_collect_data:
-            mock_collect_data.side_effect = KeyError
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-                mock_current_task.return_value = self.current_task
-
-                self.assertEqual(push_course_forums_data_to_s3(None, None, self.course.id, None, 'generated'), UPDATE_STATUS_FAILED)
-
-    @patch('lms.djangoapps.instructor_task.tasks_helper.datetime')
-    def test_report_stores_results(self, mock_time):
-        start_time = datetime.now(UTC)
-        mock_time.now.return_value = start_time
-
-        test_header = ['Date', 'Type', 'Number', 'Up Votes', 'Down Votes', 'Net Points']
-        test_rows = [['row1_field1', 'row1_field2'], ['row2_field1', 'row2_field2']]
-
-        with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-            mock_current_task.return_value = self.current_task
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper.collect_course_forums_data') as mock_collect_data:
-                mock_collect_data.return_value = (test_header, test_rows)
-
-                with patch('lms.djangoapps.instructor_task.models.DjangoStorageReportStore.store_rows'):
-                    return_val = push_course_forums_data_to_s3(None, None, self.course.id, None, 'generated')
-                    self.assertEqual(return_val, UPDATE_STATUS_SUCCEEDED)
-
-
-class TestInstructorStudentForumsReport(TestReportMixin, InstructorTaskCourseTestCase):
-    """
-    Tests that Student forums usage report generation works.
-    """
-    def setUp(self):
-        super(TestInstructorStudentForumsReport, self).setUp()
-        self.course = CourseFactory.create(org=TEST_COURSE_ORG,
-                                           number=TEST_COURSE_NUMBER,
-                                           display_name=TEST_COURSE_NAME)
-
-        self.current_task = Mock()
-        self.current_task.update_state = Mock()
-
-    def test_report_fails_if_error(self):
-        with patch('lms.djangoapps.instructor_task.tasks_helper.collect_student_forums_data') as mock_collect_data:
-            mock_collect_data.side_effect = KeyError
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-                mock_current_task.return_value = self.current_task
-
-                self.assertEqual(push_student_forums_data_to_s3(None, None, self.course.id, None, 'generated'), UPDATE_STATUS_FAILED)
-
-    @patch('lms.djangoapps.instructor_task.tasks_helper.datetime')
-    def test_report_stores_results(self, mock_time):
-        start_time = datetime.now(UTC)
-        mock_time.now.return_value = start_time
-
-        test_header = ['Username', 'Posts', 'Votes']
-        test_rows = [['row1_field1', 'row1_field2', 'row1_field3'], ['row2_field1', 'row2_field2', 'row2_field3']]
-
-        with patch('lms.djangoapps.instructor_task.tasks_helper._get_current_task') as mock_current_task:
-            mock_current_task.return_value = self.current_task
-
-            with patch('lms.djangoapps.instructor_task.tasks_helper.collect_student_forums_data') as mock_collect_data:
-                mock_collect_data.return_value = (test_header, test_rows)
-
-                with patch('lms.djangoapps.instructor_task.models.DjangoStorageReportStore.store_rows'):
-                    return_val = push_student_forums_data_to_s3(None, None, self.course.id, None, 'generated')
-                    self.assertEqual(return_val, UPDATE_STATUS_SUCCEEDED)
 
 
 @ddt.ddt
