+++ openedx/stanford/lms/djangoapps/instructor/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openedx.stanford.lms.djangoapps.instructor.apps.StanfordInstructorConfig'
