+++ openedx/stanford/lms/djangoapps/instructor/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class StanfordInstructorConfig(AppConfig):
+    name = 'openedx.stanford.lms.djangoapps.instructor'
+    label = 'instructor.stanford'
