+++ openedx/stanford/lms/djangoapps/instructor/tests/test_lti_grader.py
@@ -1,14 +1,19 @@
 """
 Tests for the Mgmt Commands Feature on the Sysadmin page
 """
-import unittest
 from mock import patch
+import unittest
 
 from django.conf import settings
 from django.core.files.uploadedfile import SimpleUploadedFile
 
 from dashboard.tests.test_sysadmin import SysadminBaseTestCase
-from instructor.lti_grader import LTIGrader
+
+from openedx.stanford.lms.djangoapps.instructor.lti_grader import LTIGrader
+
+
+always_true = lambda *args, **kwargs: True
+always_false = lambda *args, **kwargs: False
 
 
 @unittest.skipUnless(
@@ -30,9 +35,11 @@ class TestLtiGrader(SysadminBaseTestCase):
         }
         self.data = SimpleUploadedFile(
             'file.csv',
-            'ID,Anonymized User ID,email,grade, max_grade, comments,'
-            '\n1,abcdabcd,abcd@abcd.com,5,10,not bad'
-            '\n2,cdefcdef,cdef@cdef.com,6,10,great'
+            '\n'.join([
+                'ID,Anonymized User ID,email,grade, max_grade, comments,',
+                '1,abcdabcd,abcd@abcd.com,5,10,not bad',
+                '2,cdefcdef,cdef@cdef.com,6,10,great',
+            ])
         )
         self.grader = LTIGrader('course_id', 'url_base', 'lti_key', 'lti_secret')
 
@@ -52,51 +59,34 @@ class TestLtiGrader(SysadminBaseTestCase):
         self.assertEquals((2, u'cdefcdef', u'cdef@cdef.com', 6.0, 10.0, u'great'), valid_rows.next())
 
     def test_update_grades_passport_failure(self):
-        def validate_passport_side_effect(key, secret, test_url):  # pylint: disable=unused-argument
-            """
-            Side effect designed to replace validate_lti_passport in lti_connection.py, mocking validation failure
-            """
-            return False
-        with patch('instructor.lti_grader.lti_connection') as mock_lti_connection:
-            mock_lti_connection.validate_lti_passport.side_effect = validate_passport_side_effect
+        with patch('openedx.stanford.lms.djangoapps.instructor.lti_grader.lti_connection') as mock_lti_connection:
+            mock_lti_connection.validate_lti_passport.side_effect = always_false
             actual_output = self.grader.update_grades(self.data)['error']
             expected_output = ['LTI passport sanity check failed. Your lti_key (lti_key) or lti_secret (lti_secret) are probably incorrect.']
             self.assertEquals(expected_output, actual_output)
 
     def test_update_grades_success(self):
-        def validate_passport_side_effect(key, secret, test_url):  # pylint: disable=unused-argument
-            """
-            Side effect designed to replace validate_lti_passport in lti_connection.py, mocking validation success
-            """
-            return True
-
-        def post_success_side_effect(url_base, key, secret, grade_row):  # pylint: disable=unused-argument
-            """
-            Side effect designed to replace post_grade in lti_connection.py, mocking successful grade post
-            """
+        def side_effect(_url_base, _key, _secret, grade_row):
             return (True, grade_row[0], grade_row[2])
-        with patch('instructor.lti_grader.lti_connection') as mock_lti_connection:
-            mock_lti_connection.validate_lti_passport.side_effect = validate_passport_side_effect
-            mock_lti_connection.post_grade.side_effect = post_success_side_effect
+        with patch('openedx.stanford.lms.djangoapps.instructor.lti_grader.lti_connection') as mock_lti_connection:
+            mock_lti_connection.validate_lti_passport.side_effect = always_true
+            mock_lti_connection.post_grade.side_effect = side_effect
             actual_output = self.grader.update_grades(self.data)['success']
-            expected_output = ['Grade post successful: user id 1 (email: abcd@abcd.com).', 'Grade post successful: user id 2 (email: cdef@cdef.com).']
+            expected_output = [
+                'Grade post successful: user id 1 (email: abcd@abcd.com).',
+                'Grade post successful: user id 2 (email: cdef@cdef.com).',
+            ]
             self.assertEquals(expected_output, actual_output)
 
     def test_update_grades_failure(self):
-        def validate_passport_side_effect(key, secret, test_url):  # pylint: disable=unused-argument
-            """
-            Side effect designed to replace validate_lti_passport in lti_connection.py, mocking validation success
-            """
-            return True
-
-        def post_failure_side_effect(url_base, key, secret, grade_row):  # pylint: disable=unused-argument
-            """
-            Side effect designed to replace post_grade in lti_connection.py, mocking unsuccessful grade post
-            """
+        def side_effect(_url_base, _key, _secret, grade_row):
             return (False, grade_row[0], grade_row[2])
-        with patch('instructor.lti_grader.lti_connection') as mock_lti_connection:
-            mock_lti_connection.validate_lti_passport.side_effect = validate_passport_side_effect
-            mock_lti_connection.post_grade.side_effect = post_failure_side_effect
+        with patch('openedx.stanford.lms.djangoapps.instructor.lti_grader.lti_connection') as mock_lti_connection:
+            mock_lti_connection.validate_lti_passport.side_effect = always_true
+            mock_lti_connection.post_grade.side_effect = side_effect
             actual_output = self.grader.update_grades(self.data)['error']
-            expected_output = ['Grade post failed: user id 1 (email: abcd@abcd.com).', 'Grade post failed: user id 2 (email: cdef@cdef.com).']
+            expected_output = [
+                'Grade post failed: user id 1 (email: abcd@abcd.com).',
+                'Grade post failed: user id 2 (email: cdef@cdef.com).',
+            ]
             self.assertEquals(expected_output, actual_output)
