+++ openedx/stanford/lms/envs/common.py
@@ -67,6 +67,7 @@ INLINE_ANALYTICS_SUPPORTED_TYPES = {
 }
 INSTALLED_APPS += (
     'branding_stanford',
+    'openedx.stanford.lms.djangoapps.instructor',
     'settings_context_processor',
     'sneakpeek_deeplink',
     # Added here to allow translations
